backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
redis==5.0.7
rq==1.16.2
python-dateutil==2.9.0.post0
google-generativeai==0.3.2
